DataLoading
This commit completes the data loading functionality. in the App component there's some code that listens for history events. When it recevies one it fires off a redux action. That action makes a fetch the same endpoint as the server side render uses and gets some data back. From there it's just a case of rehydrating it in to the Redux store. This implementation is as bare-bones as possible and could probably so with some improvements.