Fix node app erroring on startup
When the node app started up it was looking for the react-loadable.json file. Because the node app and webpack are run in parrallel the json file had not been generated in time to be ready for the node app. This commit ensures that the json is in place before the node app starts. It's a workaround but currently I don't have any better ideas...